refactor(vfs): 将文件系统相关系统调用拆分为独立模块
将 `close`、`fstat`、`lstat` 和 `stat` 系统调用从 `mod.rs` 中拆分为独立的模块
