+++ kernel/src/filesystem/vfs/syscall/sys_stat.rs
@@ -1,0 +1,70 @@
+//! System call handler for opening files.
+
+use system_error::SystemError;
+
+use defer::defer;
+
+use crate::arch::syscall::nr::SYS_STAT;
+use crate::filesystem::vfs::file::FileMode;
+use crate::filesystem::vfs::stat::PosixKstat;
+use crate::filesystem::vfs::syscall::sys_close::do_close;
+use crate::filesystem::vfs::syscall::sys_fstat::do_fstat;
+use crate::filesystem::vfs::syscall::sys_open;
+use crate::filesystem::vfs::ModeType;
+use crate::syscall::table::FormattedSyscallParam;
+use crate::syscall::table::Syscall;
+use crate::syscall::user_access::UserBufferWriter;
+
+use alloc::vec::Vec;
+
+pub struct SysStatHandle;
+
+impl Syscall for SysStatHandle {
+    /// Returns the number of arguments this syscall takes.
+    fn num_args(&self) -> usize {
+        2
+    }
+
+    fn handle(&self, args: &[usize], _from_user: bool) -> Result<usize, SystemError> {
+        let path = Self::path(args);
+        let usr_kstat = Self::usr_kstat(args);
+        let mut writer = UserBufferWriter::new(usr_kstat, size_of::<PosixKstat>(), true)?;
+
+        let fd = sys_open::do_open(
+            path,
+            FileMode::O_RDONLY.bits(),
+            ModeType::empty().bits(),
+            true,
+        )?;
+
+        defer!({
+            do_close(fd as i32).ok();
+        });
+        let kstat = do_fstat(fd as i32)?;
+
+        writer.copy_one_to_user(&kstat, 0)?;
+        return Ok(0);
+    }
+
+    /// Formats the syscall arguments for display/debugging purposes.
+    fn entry_format(&self, args: &[usize]) -> Vec<FormattedSyscallParam> {
+        vec![
+            FormattedSyscallParam::new("path", format!("{:#x}", Self::path(args) as usize)),
+            FormattedSyscallParam::new("statbuf", format!("{:#x}", Self::usr_kstat(args) as usize)),
+        ]
+    }
+}
+
+impl SysStatHandle {
+    /// Extracts the path argument from syscall parameters.
+    fn path(args: &[usize]) -> *const u8 {
+        args[0] as *const u8
+    }
+
+    /// Extracts the usr_kstat argument from syscall parameters.
+    fn usr_kstat(args: &[usize]) -> *mut PosixKstat {
+        args[1] as *mut PosixKstat
+    }
+}
+
+syscall_table_macros::declare_syscall!(SYS_STAT, SysStatHandle);
